Move EventManager to new event package
