+++ client.go
@@ -287,10 +287,6 @@ func (c *Client) CustomizationSpecManager() CustomizationSpecManager {
 	return CustomizationSpecManager{c}
 }
 
-func (c *Client) EventManager() EventManager {
-	return EventManager{c}
-}
-
 // func (c *Client) SessionManager() SessionManager {
 // 	return NewSessionManager(c, *c.ServiceContent.SessionManager)
 // }
