+++ event/manager.go
@@ -14,39 +14,53 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package govmomi
+package event
 
 import (
+	"github.com/vmware/govmomi"
 	"github.com/vmware/govmomi/vim25/methods"
 	"github.com/vmware/govmomi/vim25/types"
 	"golang.org/x/net/context"
 )
 
-type EventManager struct {
-	c *Client
+type Manager struct {
+	reference types.ManagedObjectReference
+
+	c *govmomi.Client
 }
 
-func (em EventManager) CreateCollectorForEvents(filter types.EventFilterSpec) (*types.ManagedObjectReference, error) {
+func NewManager(c *govmomi.Client) *Manager {
+	m := Manager{
+		reference: *c.ServiceContent.EventManager,
+
+		c: c,
+	}
+
+	return &m
+}
+
+func (m Manager) CreateCollectorForEvents(ctx context.Context, filter types.EventFilterSpec) (*types.ManagedObjectReference, error) {
 	req := types.CreateCollectorForEvents{
-		This:   *em.c.ServiceContent.EventManager,
+		This:   m.reference,
 		Filter: filter,
 	}
 
-	res, err := methods.CreateCollectorForEvents(context.TODO(), em.c, &req)
+	res, err := methods.CreateCollectorForEvents(ctx, m.c, &req)
 	if err != nil {
 		return nil, err
 	}
+
 	return &res.Returnval, nil
 }
 
-func (em EventManager) LogUserEvent(entity types.ManagedObjectReference, msg string) error {
+func (m Manager) LogUserEvent(ctx context.Context, entity types.ManagedObjectReference, msg string) error {
 	req := types.LogUserEvent{
-		This:   *em.c.ServiceContent.EventManager,
+		This:   m.reference,
 		Entity: entity,
 		Msg:    msg,
 	}
 
-	_, err := methods.LogUserEvent(context.TODO(), em.c, &req)
+	_, err := methods.LogUserEvent(ctx, m.c, &req)
 	if err != nil {
 		return err
 	}
@@ -54,14 +68,14 @@ func (em EventManager) LogUserEvent(entity types.ManagedObjectReference, msg str
 	return nil
 }
 
-func (em EventManager) PostEvent(eventToPost types.BaseEvent, taskInfo types.TaskInfo) error {
+func (m Manager) PostEvent(ctx context.Context, eventToPost types.BaseEvent, taskInfo types.TaskInfo) error {
 	req := types.PostEvent{
-		This:        *em.c.ServiceContent.EventManager,
+		This:        m.reference,
 		EventToPost: eventToPost,
 		TaskInfo:    &taskInfo,
 	}
 
-	_, err := methods.PostEvent(context.TODO(), em.c, &req)
+	_, err := methods.PostEvent(ctx, m.c, &req)
 	if err != nil {
 		return err
 	}
@@ -69,29 +83,30 @@ func (em EventManager) PostEvent(eventToPost types.BaseEvent, taskInfo types.Tas
 	return nil
 }
 
-func (em EventManager) QueryEvents(filter types.EventFilterSpec) ([]types.BaseEvent, error) {
+func (m Manager) QueryEvents(ctx context.Context, filter types.EventFilterSpec) ([]types.BaseEvent, error) {
 	req := types.QueryEvents{
-		This:   *em.c.ServiceContent.EventManager,
+		This:   m.reference,
 		Filter: filter,
 	}
 
-	res, err := methods.QueryEvents(context.TODO(), em.c, &req)
+	res, err := methods.QueryEvents(ctx, m.c, &req)
 	if err != nil {
 		return nil, err
 	}
+
 	return res.Returnval, nil
 }
 
-func (em EventManager) RetrieveArgumentDescription(eventTypeID string) ([]types.EventArgDesc, error) {
+func (m Manager) RetrieveArgumentDescription(ctx context.Context, eventTypeID string) ([]types.EventArgDesc, error) {
 	req := types.RetrieveArgumentDescription{
-		This:        *em.c.ServiceContent.EventManager,
+		This:        m.reference,
 		EventTypeId: eventTypeID,
 	}
 
-	res, err := methods.RetrieveArgumentDescription(context.TODO(), em.c, &req)
+	res, err := methods.RetrieveArgumentDescription(ctx, m.c, &req)
 	if err != nil {
 		return nil, err
 	}
-	return res.Returnval, nil
 
+	return res.Returnval, nil
 }
